[TASK] Remove superfluous files from Git repository
Files of 3rd party packages that are stored in their Git repositories
do not need to be part of this project repository. Besides that temporary
files are removed from the repository as well - these can be recreated on
local installations automatically.

After cloning remote Git repository contents, `composer install` has to
be executed in order to install TYPO3 extensions and 3rd party PHP
packages defined in root composer.json.